requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
cssselect==1.3.0
datasets==4.2.0
ddgs==9.6.0
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-genai==1.11.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
multiprocess==0.70.16
nltk==3.9.2
numpy==2.3.3
openai==2.2.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
playwright==1.55.0
postgrest==2.22.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
readability-lxml==0.8.4.1
realtime==2.22.0
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
razorpay==2.0.1
rsa==4.9.1
sentry-sdk==2.40.0
setuptools>=70.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
Wikipedia-API==0.8.1
xxhash==3.6.0
yarl==1.22.0
